refactor(ui): update import paths for shared components and pages
- Update App.jsx to import moved pages from shared locations (AdminRides, AdminUsers, SignIn, Signup, ForgotPassword, Chat, VerifyEmail, EditProfile, AdminPlaceManager)
- Update MapView.jsx to use correct mobile utils path
- Update TestImageUpload.jsx to use shared component imports (ImageUpload, ImageViewer)
- Update SignIn.jsx Captcha import path (already correct, verified)
- Update Onboarding.jsx to use shared Captcha component
- Update ImRiding.jsx, MyRides.jsx, ImDriving.jsx to use shared ConfirmFunction component
- Update ComponentsTest.jsx to use shared MapView component
- All import paths now correctly reference shared components and pages
