+++ app/imports/ui/mobile/pages/MyRides.jsx
@@ -6,7 +6,7 @@ import PropTypes from "prop-types";
 import { Rides } from "../../../api/ride/Rides";
 import JoinRideModal from "../components/JoinRideModal";
 import MobileRide from "../components/Ride";
-import ConfirmFunction from "../components/ConfirmFunction";
+import ConfirmFunction from "../../components/ConfirmFunction";
 import "../../../api/chat/ChatMethods";
 import {
   Container,
